+++ docs/change_log.md
@@ -102,6 +102,7 @@ There is one significant exception to the above rules -- `MAJOR=0` releases. Any
 - added `startup_method` hook to simmate workers
 - added `baderkit` app
 - added `simmate database start` and `stop` commands for managing a Postgres database via Docker
+- added `TableCount` caching for `data_explorer` homepage counts
 
 **Refactors**